refactor(dev): split inspector presentation from raw OTLP storage; stream HTTP bodies
TraceStore becomes raw per-trace persistence; frontend shaping (meta, detail,
noise filtering) moves to core/dev/inspector with an InspectorTraceSource
query wrapper. io/httpServer responses can now stream AsyncIterable bodies
for SSE.
